move assets (#109)
