+++ pc/src/tiles.rs
@@ -1,123 +1,10 @@
+use crate::assets::{GetTexture, TEXTURES};
 pub(crate) use m3_map::tiles::MapBaseTile;
 use m3_map::{
 	story::{Background, Character},
 	tiles::{ObjectTile, PlayerTile, Tile}
 };
-use macroquad::{prelude::ImageFormat, texture::Texture2D};
-use once_cell::sync::Lazy;
-
-#[allow(clippy::redundant_closure)] //false positive?
-pub(crate) static TEXTURES: Lazy<Textures> = Lazy::new(|| Textures::init());
-
-pub(crate) trait GetTexture {
-	fn texture(&self) -> Texture2D;
-}
-
-///Store all Textures
-pub(crate) struct Textures {
-	grass: Texture2D,
-	grass_corner_sand: Texture2D,
-	half_grass_sand: Texture2D,
-	sand: Texture2D,
-	sand_corner_grass: Texture2D,
-	stone: Texture2D,
-	player1_car: Texture2D,
-	player2_car: Texture2D,
-	player3_car: Texture2D,
-	player4_car: Texture2D,
-	global_goal: Texture2D,
-	player1_goal: Texture2D,
-	player2_goal: Texture2D,
-	player3_goal: Texture2D,
-	player4_goal: Texture2D,
-	captain: Texture2D,
-	outer_space: Texture2D
-}
-
-impl Textures {
-	///init all Textures
-	pub(crate) fn init() -> Textures {
-		Textures {
-			grass: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/BaseTiles/grass.png"),
-				Some(ImageFormat::Png)
-			),
-			grass_corner_sand: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/BaseTiles/grass_corner_sand.png"),
-				Some(ImageFormat::Png)
-			),
-			half_grass_sand: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/BaseTiles/half_grass_sand.png"),
-				Some(ImageFormat::Png)
-			),
-			sand: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/BaseTiles/sand.png"),
-				Some(ImageFormat::Png)
-			),
-			sand_corner_grass: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/BaseTiles/sand_corner_grass.png"),
-				Some(ImageFormat::Png)
-			),
-			stone: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/ObjectTiles/stone.png"),
-				Some(ImageFormat::Png)
-			),
-			player1_car: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Player/player1_car.png"),
-				Some(ImageFormat::Png)
-			),
-			player2_car: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Player/player2_car.png"),
-				Some(ImageFormat::Png)
-			),
-			player3_car: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Player/player3_car.png"),
-				Some(ImageFormat::Png)
-			),
-			player4_car: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Player/player4_car.png"),
-				Some(ImageFormat::Png)
-			),
-			global_goal: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Player/goal.png"),
-				Some(ImageFormat::Png)
-			),
-			player1_goal: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Player/player1_goal.png"),
-				Some(ImageFormat::Png)
-			),
-			player2_goal: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Player/player2_goal.png"),
-				Some(ImageFormat::Png)
-			),
-			player3_goal: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Player/player3_goal.png"),
-				Some(ImageFormat::Png)
-			),
-			player4_goal: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Player/player4_goal.png"),
-				Some(ImageFormat::Png)
-			),
-			captain: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Character/captain.png"),
-				Some(ImageFormat::Png)
-			),
-			outer_space: Texture2D::from_file_with_format(
-				include_bytes!("../assets/img/Backgrounds/outer_space.png"),
-				Some(ImageFormat::Png)
-			)
-		}
-	}
-
-	pub(crate) fn get_player_textures(&self) -> Vec<Texture2D> {
-		vec![
-			self.player1_car,
-			self.player2_car,
-			self.player3_car,
-			self.player4_car,
-		]
-	}
-}
+use macroquad::texture::Texture2D;
 
 impl GetTexture for MapBaseTile {
 	///get Texture assioated with this Tile
